pllanitt/finalyearproject/backend/python/requirements.txt
fastapi
uvicorn
requests
rasterio
shapely
pyproj
python-multipart
gunicorn
numpy
scipy
matplotlib
scikit-learn
joblib
reportlab
pandas
jinja2
psycopg2-binary
openai>=1.0.0
python-dotenv
